+++ packages/liveblocks-react-comments/src/components/Composer.tsx
@@ -11,6 +11,7 @@ import type {
 } from "react";
 import React, { forwardRef, useCallback } from "react";
 
+import { EmojiIcon } from "../icons/Emoji";
 import { MentionIcon } from "../icons/Mention";
 import { SendIcon } from "../icons/Send";
 import { type ComposerOverrides, useOverrides } from "../overrides";
@@ -30,6 +31,7 @@ import { useControllableState } from "../utils/use-controllable-state";
 import { Attribution } from "./internal/Attribution";
 import { Avatar } from "./internal/Avatar";
 import { Button } from "./internal/Button";
+import { EmojiPicker, EmojiPickerTrigger } from "./internal/EmojiPicker";
 import {
   ShortcutTooltip,
   ShortcutTooltipKey,
@@ -166,6 +168,35 @@ function ComposerInsertMentionEditorAction({
         <MentionIcon className="lb-button-icon" />
       </Button>
     </Tooltip>
+  );
+}
+
+function ComposerInsertEmojiEditorAction({
+  label,
+  className,
+  ...props
+}: EditorActionProps) {
+  const { insertText } = useComposer();
+
+  const preventDefault = useCallback((event: SyntheticEvent) => {
+    event.preventDefault();
+  }, []);
+
+  return (
+    <EmojiPicker onEmojiSelect={insertText}>
+      <Tooltip content={label}>
+        <EmojiPickerTrigger asChild>
+          <Button
+            className={classNames("lb-composer-editor-action", className)}
+            onMouseDown={preventDefault}
+            aria-label={label}
+            {...props}
+          >
+            <EmojiIcon className="lb-button-icon" />
+          </Button>
+        </EmojiPickerTrigger>
+      </Tooltip>
+    </EmojiPicker>
   );
 }
 
@@ -335,6 +366,9 @@ const ComposerWithContext = forwardRef<
                   label={$.COMPOSER_INSERT_MENTION}
                 />
               )}
+              <ComposerInsertEmojiEditorAction
+                label={$.COMPOSER_INSERT_EMOJI}
+              />
             </div>
             {showAttribution && <Attribution />}
             <div className="lb-composer-actions">