+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - **Breaking (beta):** `Comment`’s `indentBody` and `Thread`’s
   `indentCommentBody` were renamed to `indentContent` and `indentCommentContent`
   respectively. `Thread`’s `onResolveChange` was renamed to `onResolvedChange`.
+- Add emoji button in `Composer`.
 - Broadcast event messages now include a `user` property to indicate the user
   that sent the event:
   ```tsx